# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/admin/sessions/{sessionId}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Позволяет администратору группы изменить данные сессии, принадлежащей этой группе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сессии_админ"
                ],
                "summary": "Обновить информацию о сессии",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID сессии",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные сессии",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.SessionUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сессия успешно обновлена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации или некорректный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Нет прав на редактирование",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Сессия не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/sessions/createSession": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает сессию в группе с возможностью загрузки изображения и добавления метаданных",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сессии"
                ],
                "summary": "Создание новой сессии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название сессии",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Тип сессии",
                        "name": "session_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Типа проведения сессии",
                        "name": "session_place",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "group_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Время начала (в формате RFC3339, напр. 2025-07-10T19:00:00+02:00)",
                        "name": "start_time",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Длительность в минутах",
                        "name": "duration",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Максимальное количество участников",
                        "name": "count_users",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Жанры (через запятую, напр: драма,комедия)",
                        "name": "genres",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Доп. поля (напр: ключ:значение,ключ2:знач2)",
                        "name": "fields",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Место проведения",
                        "name": "location",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Год (например: 2023)",
                        "name": "year",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Страна",
                        "name": "country",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Возрастное ограничение (напр: 16+)",
                        "name": "age_limit",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Примечания",
                        "name": "notes",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Изображение",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сессия создана",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не передан JWT",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера или валидации",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/sessions/genres": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех доступных жанров для сессий.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сессии"
                ],
                "summary": "Получение списка жанров",
                "responses": {
                    "200": {
                        "description": "Список жанров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/sessions/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Позволяет пользователю присоединиться к выбранной сессии, если она не заполнена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сессии"
                ],
                "summary": "Присоединение к сессии",
                "parameters": [
                    {
                        "description": "Данные для присоединения к группе",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SessionJoinInputDoc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вы успешно присоединились к сессии",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка разбора формы",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Сессия или пользователь не найдены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Сессия заполнена или пользователь уже присоединился",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/sessions/sessions/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет сессию, если пользователь является admin или operator в группе",
                "tags": [
                    "Сессии"
                ],
                "summary": "Удаление сессии",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID сессии для удаления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сессия удалена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "JWT не передан",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Сессия не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера или БД",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/sessions/{sessionId}/leave": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Позволяет пользователю покинуть сессию, в которой он участвует",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сессии"
                ],
                "summary": "Отписаться от сессии",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID сессии",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вы успешно покинули сессию",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Вы не состоите в сессии или нет доступа",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Сессия не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.SessionJoinInputDoc": {
            "type": "object",
            "required": [
                "group_id",
                "session_id"
            ],
            "properties": {
                "group_id": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "integer"
                }
            }
        },
...
        "services.SessionUpdateInput": {
            "type": "object",
            "properties": {
                "count_users_max": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer \u003cyour_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}